pyproject.toml
[project]
name = "docrefinery"
version = "0.1.0"
description = "The Document Intelligence Refinery – multi-stage agentic pipeline for structured, provenance-preserving document extraction."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "FDE", email = "fde@example.com" }
]

dependencies = [
  "pydantic>=2.6.0",
  "pdfplumber>=0.11.0",
  "docling>=2.0.0",
  "langdetect>=1.0.9",
  "pyyaml>=6.0.1",
   "python-dotenv>=1.0.1",
  "chromadb>=0.5.0",
  "sqlalchemy>=2.0.0",
  "httpx>=0.27.0",
  "langgraph>=0.2.0",
  "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=4.1.0",
  "mypy>=1.8.0",
  "ruff>=0.5.0",
  "reportlab>=4.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

